requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
#av==10.0.0
babel==2.17.0
bangla==0.0.5
basicsr==1.4.2
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cloudpathlib==0.22.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==4.4.2
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
email-validator==2.3.0
encodec==0.1.1
exceptiongroup==1.3.0
face-alignment==1.3.5
facexlib==0.3.0
fastapi==0.111.1
fastapi-cli==0.0.10
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.18.0
filterpy==1.4.5
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.5.0
future==1.0.0
g2pkk==0.1.2
gast==0.4.0
gfpgan==1.3.8
google-auth==2.40.2
google-auth-oauthlib==1.0.0
gradio==3.41.2
gradio_client==0.5.0
grpcio==1.74.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.4
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib_resources==6.5.2
#inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kornia==0.6.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0
llvmlite==0.43.0
lmdb==1.7.3
lxml==6.0.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
narwhals==2.3.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
orjson==3.11.3
packaging==25.0
pandas==1.5.3
pdf2image==1.17.0
Pillow==10.0.0
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.10
proglog==0.1.12
propcache==0.3.2
protobuf==4.25.7
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.55.0
pysbd==0.3.4
pytesseract==0.3.13
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ffmpeg==2.0.12
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.3.1
resnet==0.1
rich==14.1.0
rich-toolkit==0.15.1
rpds-py==0.27.1
rsa==4.9.1
ruff==0.12.12
safetensors==0.6.2
scikit-image==0.19.3
scikit-learn==1.7.1
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.37.2
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.14.0
tb-nightly==2.21.0a20250907
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.22.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.15.2
tomli==2.2.1
tomlkit==0.12.0
# torch==2.1.2+cu118
# torchaudio==2.1.2+cu118
# torchvision==0.16.2+cu118
tqdm==4.67.1
trainer==0.0.36
transformers==4.35.2
#triton==2.1.0
TTS==0.22.0
typeguard==2.13.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.30.1
#uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.14.1
xlsxwriter==3.2.5
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1